+++ Marlin/UBL_Bed_Leveling.cpp
@@ -59,7 +59,7 @@
 
   ubl_state unified_bed_leveling::state, unified_bed_leveling::pre_initialized;
 
-  float unified_bed_leveling::z_values[UBL_MESH_NUM_X_POINTS][UBL_MESH_NUM_Y_POINTS],
+  double unified_bed_leveling::z_values[UBL_MESH_NUM_X_POINTS][UBL_MESH_NUM_Y_POINTS],
         unified_bed_leveling::last_specified_z,
         unified_bed_leveling::mesh_index_to_xpos[UBL_MESH_NUM_X_POINTS + 1], // +1 safety margin for now, until determinism prevails
         unified_bed_leveling::mesh_index_to_ypos[UBL_MESH_NUM_Y_POINTS + 1];
@@ -101,7 +101,7 @@
        * updated, but until then, we try to ease the transition
        * for our Beta testers.
        */
-      const float recip = ubl.state.g29_correction_fade_height ? 1.0 / ubl.state.g29_correction_fade_height : 1.0;
+      const double recip = ubl.state.g29_correction_fade_height ? 1.0 / ubl.state.g29_correction_fade_height : 1.0;
       if (ubl.state.g29_fade_height_multiplier != recip) {
         ubl.state.g29_fade_height_multiplier = recip;
         store_state();
@@ -190,7 +190,7 @@
       SERIAL_EOL;
     }
 
-    const float current_xi = ubl.get_cell_index_x(current_position[X_AXIS] + (MESH_X_DIST) / 2.0),
+    const double current_xi = ubl.get_cell_index_x(current_position[X_AXIS] + (MESH_X_DIST) / 2.0),
                 current_yi = ubl.get_cell_index_y(current_position[Y_AXIS] + (MESH_Y_DIST) / 2.0);
 
     for (uint8_t j = UBL_MESH_NUM_Y_POINTS - 1; j >= 0; j--) {
@@ -200,7 +200,7 @@
         // is the nozzle here? then mark the number
         if (map0) SERIAL_CHAR(is_current ? '[' : ' ');
 
-        const float f = z_values[i][j];
+        const double f = z_values[i][j];
         if (isnan(f)) {
           serialprintPGM(map0 ? PSTR("   .  ") : PSTR("NAN"));
         }
