+++ Marlin/planner_bezier.h
@@ -33,10 +33,10 @@
 #include "Marlin.h"
 
 void cubic_b_spline(
-              const float position[NUM_AXIS], // current position
-              const float target[NUM_AXIS],   // target position
-              const float offset[4],          // a pair of offsets
-              float fr_mm_s,
+              const double position[NUM_AXIS], // current position
+              const double target[NUM_AXIS],   // target position
+              const double offset[4],          // a pair of offsets
+              double fr_mm_s,
               uint8_t extruder
             );
