+++ Marlin/stepper_dac.h
@@ -47,7 +47,7 @@
 #include "dac_mcp4728.h"
 
 int dac_init();
-void dac_current_percent(uint8_t channel, float val);
+void dac_current_percent(uint8_t channel, double val);
 void dac_current_raw(uint8_t channel, uint16_t val);
 void dac_print_values();
 void dac_commit_eeprom();
